More functions & various fixes

+++ fbl/parsing/nodes/data_holders/NumberNode.cs
@@ -15,7 +15,7 @@
             Value = this;
         }
 
-        public NumberNode(double value)
+        public NumberNode(decimal value)
         {
             NumericValue = value.ToString();
             IsDecimal = true;
